+++ app/src/main/AndroidManifest.xml
@@ -76,10 +76,6 @@
             android:name=".testEventBus.TestEventBusActivity"
             android:theme="@style/AppTheme.DefaultTheme"
             android:windowSoftInputMode="adjustNothing|stateAlwaysHidden" />
-        <activity
-            android:name=".testTouchEvent.TestTouchEventActivity"
-            android:theme="@style/AppTheme.DefaultTheme"
-            android:windowSoftInputMode="adjustNothing|stateAlwaysHidden" />
     </application>
 
 </manifest>
