+++ research/common/src/main/AndroidManifest.xml
@@ -12,6 +12,10 @@
             android:name=".aidl.TestAIDLActivity"
             android:theme="@style/AppTheme.DefaultTheme"
             android:windowSoftInputMode="adjustNothing|stateAlwaysHidden" />
+        <activity
+            android:name=".touch.TestTouchEventActivity"
+            android:theme="@style/AppTheme.DefaultTheme"
+            android:windowSoftInputMode="adjustNothing|stateAlwaysHidden" />
 
         <service
             android:name=".service.MessageManagerService"
