+++ research/common/src/main/java/com/dvsnier/test/common/touch/TestTouchEventActivity.java
@@ -1,10 +1,10 @@
-package com.dvsnier.testTouchEvent;
+package com.dvsnier.test.common.touch;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 
-import com.dvsnier.R;
 import com.dvsnier.base.flavor.activity.BaseActivity;
+import com.dvsnier.test.common.R;
 
 public class TestTouchEventActivity extends BaseActivity {
