+++ research/common/src/main/java/com/dvsnier/test/common/touch/touchEvent.txt
