+++ research/common/src/main/res/layout/activity_test_touch_event.xml
@@ -1,21 +1,21 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.dvsnier.view.ParentLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<com.dvsnier.test.common.ParentLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@android:color/holo_blue_dark"
-    android:paddingBottom="@dimen/activity_vertical_margin"
     android:paddingLeft="@dimen/activity_horizontal_margin"
+    android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin">
+    android:paddingBottom="@dimen/activity_vertical_margin">
 
-    <com.dvsnier.view.MiddleLayout
+    <com.dvsnier.test.common.MiddleLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:layout_margin="@dimen/activity_margin"
         android:background="@android:color/darker_gray">
 
-        <com.dvsnier.view.MyTouchView
+        <com.dvsnier.test.common.MyTouchView
             android:layout_width="100dp"
             android:layout_height="100dp"
             android:layout_centerInParent="true"
@@ -24,6 +24,6 @@
             android:gravity="center"
             android:text="地球人，您好..."
             tools:ignore="HardcodedText" />
-    </com.dvsnier.view.MiddleLayout>
+    </com.dvsnier.test.common.MiddleLayout>
 
-</com.dvsnier.view.ParentLayout>
+</com.dvsnier.test.common.ParentLayout>
